Use implicit latest revision in current Tool Shed as deps.
This change means the toolshed and repository are filled in by
the Tool Shed with the latest revision in the current Tool Shed
at the time of upload.

This allows me to upload the same file to both the Test and Main
Tool Shed, but is not ideal in that the exact version desired is
not explicit in the repository_dependencies.xml file (and may
clash with the non-revision based old-style version in the *.ga
workflow files).