Test script to prepare an Arabidopsis thaliana tRNA diagram
$ python fetch_data.py

That should download 10 files (FASTA and an annotation file for each chr),
and merge them to make a 120MB FASTA file and a 31MB tabular file.

$ python chromosome_diagram.py A_thaliana.fasta 5000 A_thaliana_tRNA.tsv 2
3 4 5 1 7 8 "Arabidopsis thaliana" 0 A_thaliana_tRNA.pdf

This should produce a single page PDF file showing the five chromosomes,
with the tRNA shown in red, green, blue (with no background/fill), and
in black with a yellow fill. There should be a single run of NNNN
marked in the middle of chrI.