CLC Bio: Use environment variable; 4 thread default; Enable tests.
Using the new repeat/conditional syntax with pipe characters
which probably only works using the test framework with the
API backend (and not the older Twill backend).

To run the tests I have to log into our server with the CLC
licence (since the test framework runs all jobs locally).