Alternative idea for reporting/checking MIRA version
My thinking is to call this in <version_command> to
report the version. Idealy I would do the check here
too, provided it can make Galaxy abort.

Otherwise, would have to prefix the <command> with calling
this script and verifying the expected version.

In cases were we still require a Python wrapper script, that
can call this script directly to avoid repetition...