Revert "TravisCI: Run linting as initial TravisCI stage"
This reverts commit 6592ff977b79b363d8ac132ed18aedc21b30e4bb.

This worked, but the time spent linting (about 10s) is dwarfed by
the total overhead of a TravisCI build (over a minute, including
up to 30s just to install planemo and its dependencies).

The time taken to install flake8 and restructuredtext-lint is
minor.

Overall it is better to do the linting at the start of each
test build (which already have to install planemo anyway).