Python 3 iteritems fix
The usage here is on a small dictionary, so there is
no hard using the list-based iteration on Python 2.

Spotted with flake8-bugbear,

B301 Python 3 does not include `.iter*` methods on dictionaries. Remove the `iter` prefix from the method name. For Python 2 compatibility, prefer the Python 3 equivalent unless you expect the size of the container to be large or unbounded. Then use `six.iter*` or `future.utils.iter*`.